Fix include order issue for gfx1250
